requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.6
SQLAlchemy==2.0.18
python-dotenv==1.0.0
requests==2.31.0
bcrypt==4.0.1
gunicorn==20.1.0
python-dateutil==2.8.2
openpyxl==3.1.2
pandas==2.0.3
matplotlib==3.7.1
seaborn==0.12.2
Pillow==10.0.0